新增字段及功能

+++ src/main/java/org/game/beatingpicture/entity/Product.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class Product {
-    private int id;
-    private int pictureId;
-    private int categoryId;
-    private int price;
-    private float discount;
-    private int status;
+    private Integer id;
+    private Integer pictureId;
+    private Integer categoryId;
+    private Integer price;
+    private Integer discount;
+    private Integer status;
 
 }
