+++ src/main/java/org/game/beatingpicture/service/impl/ExpServiceImpl.java
@@ -1,5 +1,6 @@
 package org.game.beatingpicture.service.impl;
 
+import io.swagger.models.auth.In;
 import org.game.beatingpicture.mapper.ExpMapper;
 import org.game.beatingpicture.service.ExpService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,22 +13,22 @@ public class ExpServiceImpl implements ExpService {
     private ExpMapper expMapper;
 
     @Override
-    public long getExpByRankId(int rankId) {
+    public Integer getExpByRankId(Integer rankId) {
         return expMapper.getExpByRankId(rankId);
     }
 
     @Override
-    public void updateRankByUserInfoId(int userInfoId) {
+    public void updateRankByUserInfoId(Integer userInfoId) {
         expMapper.updateRankByUserInfoId(userInfoId);
     }
 
     @Override
-    public void updateExpByUserInfoId(long gExp, int userInfoId) {
+    public void updateExpByUserInfoId(Integer gExp, Integer userInfoId) {
         expMapper.updateExpByUserInfoId(gExp,userInfoId);
     }
 
     @Override
-    public void updateExpByUserInfoIdSet(long gExp, int userInfoId) {
+    public void updateExpByUserInfoIdSet(Integer gExp, Integer userInfoId) {
         expMapper.updateExpByUserInfoIdSet(gExp,userInfoId);
     }
 }
